Allows user to display love message spanning more than 5 lines
The purpose of this commit is to remove restrictions on how long a message can be.
The user can now enter/use a love message of any length and it will be displayed properly.
Up to the first five lines are still printed as normal. Anything past that gets printed
below the heart, but still aligned with the rest of the text.

format_quote() and main() were updated to support this change.