Allows user to display love message spanning more than 5 lines (#15)
* Allows user to display love message spanning more than 5 lines

The purpose of this commit is to remove restrictions on how long a message can be.
The user can now enter/use a love message of any length and it will be displayed properly.
Up to the first five lines are still printed as normal. Anything past that gets printed
below the heart, but still aligned with the rest of the text.

format_quote() and main() were updated to support this change.

* Fixed index out of bounds issue

Added try / except to catch index of bounds error when creating quoteList.
This issue arrised when code was modified to allow users to display
love messages of any length. Everything should work successfully now.

If index goes out of bounds while populating quoteList, it will apend
the element instead of using trying to access an index that doesn't exist.